sync should be working

+++ .local/bin/music-sync.sh
@@ -1,23 +1,22 @@
 #!/bin/bash
 # UPLOAD NEW MUSIC
 mkdir -p ~/Downloads/new-music
-rclone copy ~/Downloads/new-music gdrive:Music
-if [ $? -eq 0 ]; then
-    rm -f ~/Downloads/new-music/*
-    rclone dedupe --dedupe-mode newest gdrive:Music
-else
-    echo "Aborting: Failed to upload all files!"
-    exit 1
+if [ -n "$(ls ~/Downloads/new-music)" ]; then
+    rclone copy ~/Downloads/new-music gdrive:Music
+    if [ $? -eq 0 ]; then
+        rm -f ~/Downloads/new-music/*
+        rclone dedupe --dedupe-mode newest gdrive:Music
+    else
+        echo "Aborting: Failed to upload all files!"
+        exit 1
+    fi
 fi
 
-# TWO WAY SYNC
-#   set -e
-#   LEDGER_DIR="$HOME/.cache/rclone/bisync"
-#   if ! ls "$LEDGER_DIR"/gdrive_Music* &>/dev/null; then
-#       rclone bisync gdrive:Music ~/Music --resync
-#   else
-#       rclone bisync gdrive:Music ~/Music
-#   fi
+# TWO WAY SYNK DELETE > MODIFY
+rclone sync gdrive:Music ~/Music --ignore-existing
+rclone sync ~/Music gdrive:Music --ignore-existing
+rclone bisync gdrive:Music ~/Music
+
 
 # CREATE SNAPSHOT
 #   LATEST_SNAPSHOT=$(rclone lsd gdrive:Music-Snapshots/ 2>/dev/null | awk '{print $NF}' | sort | tail -1)
